refactor: 피드백반영 - mvc용 컨트롤러, rest 컨트롤러 분리
