+++ src/main/java/techcourse/myblog/Runner.java
@@ -3,29 +3,19 @@ package techcourse.myblog;
 import org.springframework.boot.ApplicationArguments;
 import org.springframework.boot.ApplicationRunner;
 import org.springframework.stereotype.Component;
-import techcourse.myblog.domain.Article;
-import techcourse.myblog.domain.User;
-import techcourse.myblog.service.ArticleService;
-import techcourse.myblog.service.CommentService;
 import techcourse.myblog.service.UserService;
-import techcourse.myblog.service.dto.ArticleRequest;
-import techcourse.myblog.service.dto.CommentRequest;
 import techcourse.myblog.service.dto.UserRequest;
 
 @Component
 public class Runner implements ApplicationRunner {
     private UserService userService;
-    private ArticleService articleService;
-    private CommentService commentService;
 
-    public Runner(UserService userService, ArticleService articleService, CommentService commentService) {
+    public Runner(UserService userService) {
         this.userService = userService;
-        this.articleService = articleService;
-        this.commentService = commentService;
     }
 
     @Override
-    public void run(ApplicationArguments args) throws Exception {
+    public void run(ApplicationArguments args) {
         UserRequest userRequest = new UserRequest();
         userRequest.setName("CU");
         userRequest.setEmail("root@gmail.com");
