+++ src/main/java/techcourse/myblog/web/ArticleController.java
@@ -2,19 +2,16 @@ package techcourse.myblog.web;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.*;
-import techcourse.myblog.domain.Article;
-import techcourse.myblog.domain.User;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 import techcourse.myblog.service.ArticleService;
-import techcourse.myblog.service.dto.ArticleRequest;
 import techcourse.myblog.service.dto.CommentResponse;
 
-import javax.validation.Valid;
 import java.util.List;
 
-@Controller
+@RestController
 @RequestMapping("/articles")
 public class ArticleController {
     private ArticleService articleService;
@@ -23,48 +20,7 @@ public class ArticleController {
         this.articleService = articleService;
     }
 
-    @GetMapping()
-    public String formArticle(Model model) {
-        model.addAttribute("article", null);
-        return "article-edit";
-    }
-
-    @PostMapping()
-    public String saveArticle(@Valid ArticleRequest articleRequest, Model model, User user) {
-        Article article = articleService.save(articleRequest, user);
-        model.addAttribute("article", article);
-        return "redirect:/articles/" + article.getId();
-    }
-
-    @GetMapping("/{articleId}")
-    public String selectArticle(@PathVariable("articleId") long articleId, Model model) {
-        Article article = articleService.findById(articleId);
-        model.addAttribute("article", article);
-        return "article";
-    }
-
-    @GetMapping("/{articleId}/edit")
-    public String edit(@PathVariable("articleId") long articleId, Model model, User user) {
-        Article article = articleService.findByIdWithUser(articleId, user);
-        model.addAttribute("article", article);
-        return "article-edit";
-    }
-
-    @PutMapping("/{articleId}")
-    public String editArticle(@PathVariable("articleId") long articleId, @ModelAttribute ArticleRequest articleRequest, User user, Model model) {
-        Article article = articleService.editArticle(articleRequest, articleId, user);
-        model.addAttribute("article", article);
-        return "redirect:/articles/" + articleId;
-    }
-
-    @DeleteMapping("/{articleId}")
-    public String deleteArticle(@PathVariable("articleId") long articleId, User user) {
-        articleService.deleteById(articleId, user);
-        return "redirect:/";
-    }
-
     @GetMapping("/{articleId}/comment")
-    @ResponseBody
     public ResponseEntity<List<CommentResponse>> comments(@PathVariable("articleId") Long articleId) {
         List<CommentResponse> commentResponses = articleService.findAllComment(articleId);
         return new ResponseEntity<>(commentResponses, HttpStatus.OK);
